SRC_Project2/Tests/requirements.txt
pyarrow
dnspython==2.6.1
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
pygeoip==0.3.2
